Refactor: Moved Command modules in their own folder
