+++ packages/core/src/classes/modules/commands/ContextMenuCommandModule.ts
@@ -1,7 +1,7 @@
 import { ContextMenuCommandBuilder } from "discord.js";
-import { CommandType } from "../../types/constants.js";
+import { CommandType } from "../../../types/constants.js";
 import { CommandModule, CommandModuleData } from "./CommandModule.js";
-import { ContextMenuCommandExecuteFunction } from "../../types/structures.js";
+import { ContextMenuCommandExecuteFunction } from "../../../types/structures.js";
 
 export interface ContextMenuCommandModuleData extends CommandModuleData {
     command_type: CommandType.ContextMenuCommand;
