+++ packages/core/src/classes/modules/commands/MessageCommandModule.ts
