+++ packages/core/src/classes/modules/commands/SlashCommandModule.ts
